// frontend/src/core/api/generated/model/productionItemRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

export interface ProductionItemRequest {
  /**
   * @minLength 0
   * @maxLength 60
   */
  trade: string
  /**
   * @minLength 0
   * @maxLength 120
   * @nullable
   */
  title?: string | null
  /** @nullable */
  crewId?: string | null
  /** @exclusiveMinimum 0 */
  totalQuantity: number
  /**
   * @minLength 0
   * @maxLength 12
   */
  unit: string
  /** @nullable */
  startDate?: string | null
  /** @nullable */
  plannedEnd?: string | null
  /**
   * @minLength 0
   * @maxLength 500
   * @nullable
   */
  note?: string | null
}